DOMA-1523 move building edit top menu to separate component
